% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_ondri_dates_to_iso_dates}
\alias{convert_ondri_dates_to_iso_dates}
\title{A date converter}
\usage{
convert_ondri_dates_to_iso_dates(date_column)
}
\arguments{
\item{date_column}{a vector presumed to be of type DATE from an ONDRI_df}
}
\description{
Convert an ONDRI date column to standard via \code{as.Date}
}
