% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_missing}
\alias{find_missing}
\title{Find and map missing data from a ONDRI_df}
\usage{
find_missing(ondridf)
}
\arguments{
\item{ondridf}{an ONDRI_df}
}
\description{
Given an ONDRI_df, this returns a data.frame (of class \code{ONDRI_df_missing}) that contains a mapping of missing values by row, column, subject (if available), and variable. Also includes the text M-code and the \code{haven} tagged NA for reference.
}
