% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ondridf_by_data_type}
\alias{get_ondridf_by_data_type}
\title{Get subset ONDRI_df by data type}
\usage{
get_ondridf_by_data_type(ondridf, data_type)
}
\arguments{
\item{ondridf}{the ONDRI_df data frame}

\item{data_type}{one of the allowable ONDRI data types (see \code{data_types()})}
}
\description{
Get the ONDRI_df of strictly the particular ONDRI data type
}
