% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ondridf_by_DATE}
\alias{get_ondridf_by_DATE}
\title{Get subset ONDRI_df by DATE}
\usage{
get_ondridf_by_DATE(ondridf)
}
\arguments{
\item{ondridf}{the ONDRI_df data frame}
}
\description{
Get the ONDRI_df susbet of only type DATE
}
