% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ondridf_by_MIXED}
\alias{get_ondridf_by_MIXED}
\title{Get subset ONDRI_df by MIXED}
\usage{
get_ondridf_by_MIXED(ondridf)
}
\arguments{
\item{ondridf}{the ONDRI_df data frame}
}
\description{
Get the ONDRI_df susbet of only type MIXED
}
