% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ondridf_column_data_types}
\alias{get_ondridf_column_data_types}
\title{Get ONDRI_df column types}
\usage{
get_ondridf_column_data_types(ondridf)
}
\arguments{
\item{ondridf}{a ONDRI_df}
}
\description{
Gets the attributes of a ONDRI_df data.frame
}
