% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ondridf_column_names_by_CATEGORICAL}
\alias{get_ondridf_column_names_by_CATEGORICAL}
\title{Get columns names by CATEGORICAL}
\usage{
get_ondridf_column_names_by_CATEGORICAL(ondridf)
}
\arguments{
\item{ondridf}{the ONDRI_df data frame}
}
\description{
Get all the column names of type CATEGORICAL
}
