% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ondridf_column_names_by_DATE}
\alias{get_ondridf_column_names_by_DATE}
\title{Get columns names by DATE}
\usage{
get_ondridf_column_names_by_DATE(ondridf)
}
\arguments{
\item{ondridf}{the ONDRI_df data frame}
}
\description{
Get all the column names of type DATE
}
