% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ondridf_column_names_by_TEXT}
\alias{get_ondridf_column_names_by_TEXT}
\title{Get columns names by TEXT}
\usage{
get_ondridf_column_names_by_TEXT(ondridf)
}
\arguments{
\item{ondridf}{the ONDRI_df data frame}
}
\description{
Get all the column names of type TEXT
}
