% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_ONDRI_df}
\alias{is_ONDRI_df}
\title{is_ONDRI_df}
\usage{
is_ONDRI_df(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class "ONDRI_df", \code{FALSE} otherwise.
}
\description{
Tests if the \code{x} object is of class type "ONDRI_df"
}
\seealso{
\code{\link{inherits}}
}
