% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_ONDRI_df_missing}
\alias{is_ONDRI_df_missing}
\title{is_ONDRI_df_missing}
\usage{
is_ONDRI_df_missing(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class "ONDRI_df_missing", \code{FALSE} otherwise.
}
\description{
Not in use yet.
}
\details{
Tests if the \code{x} object is of class type "ONDRI_df_missing"
}
\seealso{
\code{\link{inherits}}
}
