% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondridf.R
\name{ONDRI_df}
\alias{ONDRI_df}
\title{ONDRI data frame}
\usage{
ONDRI_df(DATA_file, DICTIONARY_file)
}
\arguments{
\item{DATA_file}{a character/string that is the path to an ONDRI DATA.csv file}

\item{DICTIONARY_file}{a character/string that is the path to an ONDRI DICT.csv file}
}
\description{
Provide paths to the DATA and DICT file pair and a new ONDRI_df object is returned. It is of type data.frame with some additional features including specialized NAs and printing.
}
