% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondridf.R
\docType{package}
\name{ONDRIdf-package}
\alias{ONDRIdf-package}
\alias{_PACKAGE}
\alias{ondridf}
\alias{ONDRIDF}
\alias{ONDRIdf}
\title{ONDRIdf}
\description{
ONDRIdf is designed for one general purpose: to read in and (more) easily handle ONDRI data, which has multiple types of missing data and a paired dictionary file.
}
\details{
As of now, ONDRIdf provides one primary function: \code{ONDRI_df} and multiple helpers (see the index of the help for the package).
The helpers include, for example, ways to get the indices and types of missing data, and to get the names or subset an ONDRI_df based on ONDRI data types.
}
\author{
\strong{Maintainer}: Derek Beaton \email{exposition.software@gmail.com}

}
\keyword{internal}
