% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{tagged_na_map}
\alias{tagged_na_map}
\title{Mapping of MISSING codes to \code{haven} tags and descriptors}
\usage{
tagged_na_map()
}
\description{
a function that produces a data.frame that is the set of ONDRI missing codes with all needed information
}
